cleanup and organize files
